+++ 3rdParty/libmsc/include/libmsc.h
@@ -1,5 +1,5 @@
 
-/** $VER: libmsc.h (2026.04.10) P. Stuer - My Support Classes, The "Most Original Name" Winner **/
+/** $VER: libmsc.h (2026.08.14) P. Stuer - My Support Classes, The "Most Original Name" Winner **/
 
 #pragma once
 
@@ -8,10 +8,9 @@
 #error Unicode character set compilation not enabled.
 #endif
 
-#include <SDKDDKVer.h>
-
 #define NOMINMAX
 
+#include <SDKDDKVer.h>
 #include <windows.h>
 
 #include <filesystem>
@@ -22,38 +21,8 @@ namespace fs = std::filesystem;
 #include "Encoding.h"
 #include "Enum.h"
 #include "Exception.h"
+#include "NLS.h"
 #include "RAII.h"
 #include "Stream.h"
 #include "Support.h"
-
-namespace msc
-{
-
-/// <summary>
-/// A more sane way of representing a rectangle
-/// </summary>
-struct rect_t
-{
-    rect_t & operator = (const D2D1_RECT_F & other) noexcept
-    {
-        *this = other;
-
-        return *this;
-    }
-
-    operator D2D1_RECT_F () const noexcept
-    {
-        return { x1, y1, x2, y2 };
-    }
-
-    D2D1_SIZE_F Size() const noexcept { return { std::abs(x1 - x2), std::abs(y1 - y2) }; }
-    FLOAT Width() const noexcept { return std::abs(x2 - x1); }
-    FLOAT Height() const noexcept { return std::abs(y2 - y1); }
-
-    FLOAT x1;
-    FLOAT y1;
-    FLOAT x2;
-    FLOAT y2;
-};
-
-}
+#include "Win32.h"